Move aggregators to use more specific types and allow for using dates there
TODO: add tests too

+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/SpecializedStorage.java
@@ -15,7 +15,7 @@ public abstract class SpecializedStorage<T> extends Storage {
   protected abstract T[] newUnderlyingArray(int size);
 
   @Override
-  public abstract long getType();
+  public abstract int getType();
 
   /**
    * @param data the underlying data
